Remove warnings

+++ edu.rice.cs.hpctraceviewer.ui/src/edu/rice/cs/hpctraceviewer/ui/blamestat/CpuBlameAnalysis.java
@@ -116,7 +116,7 @@ public class CpuBlameAnalysis implements IPixelAnalysis
 		gpu_idle_count   = new HashMap<>();
 		
 		cpuBlameMap.clear();		
-		cpuTotalBlame = (float) 0;
+		cpuTotalBlame = 0;
 
 	}
